+++ example/cosmosdb/src/data.ts
@@ -1,0 +1,9 @@
+import { CosmosDataSource } from "apollo-datasource-cosmosdb";
+
+export type Message = {
+  id: string;
+  text: string;
+  type: "message";
+};
+
+export class MessageDataSource extends CosmosDataSource<Message, unknown> {}