+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog for graphql-azure-subscriptions
 
+## [Unreleased]
+
+### @aaronpowell/graphql-signalr-subscriptions
+
+- Version Bump
+
+### @aaronpowell/graphql-cosmosdb-subscriptions
+
+#### Added
+
+- Initial Release
+
 ## 0.1.0 - 2021-02-17
 
 ### @aaronpowell/graphql-signalr-subscriptions